Move ClusterInfo, NodeInfo, and ReplicaInfo to org.astraea.app.admin
