+++ app/src/main/java/org/astraea/app/admin/NodeInfo.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.app.cost;
+package org.astraea.app.admin;
 
 import java.util.Objects;
