+++ app/src/test/java/org/astraea/app/admin/ClusterBeanTest.java
@@ -18,7 +18,6 @@ package org.astraea.app.admin;
 
 import java.util.List;
 import java.util.Map;
-import org.astraea.app.cost.ClusterInfo;
 import org.astraea.app.metrics.KafkaMetrics;
 import org.astraea.app.metrics.broker.HasValue;
 import org.astraea.app.metrics.jmx.BeanObject;
