+++ app/src/test/java/org/astraea/app/admin/ClusterInfoWithOfflineNodeTest.java
@@ -20,7 +20,6 @@ import java.util.Set;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
 import org.astraea.app.common.Utils;
-import org.astraea.app.cost.ReplicaInfo;
 import org.astraea.app.service.RequireBrokerCluster;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
