+++ app/src/test/java/org/astraea/app/cost/FakeClusterInfo.java
@@ -20,6 +20,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import org.astraea.app.admin.ClusterBean;
+import org.astraea.app.admin.ClusterInfo;
+import org.astraea.app.admin.NodeInfo;
+import org.astraea.app.admin.ReplicaInfo;
 
 public class FakeClusterInfo implements ClusterInfo {
   @Override
