+++ app/src/test/java/org/astraea/app/cost/NodeInfoTest.java
@@ -17,6 +17,7 @@
 package org.astraea.app.cost;
 
 import org.apache.kafka.common.Node;
+import org.astraea.app.admin.NodeInfo;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
